% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{allele_filter}
\alias{allele_filter}
\title{Determine allele assignments that are consistent with each IBD pair}
\usage{
allele_filter(al.df)
}
\arguments{
\item{al.df}{Dataframe of allele assignments (one assignment per
row) across all genotypes (columns)}
}
\value{
List (hash table) that maps pairs of genotypes (represented as a
hyphen-separated string) to row numbers of \code{al.df} that have the
pair of genotypes sharing the same allele.
}
\description{
Pre-process the allele assignments such that the enumeration required to
compte p(y|IBD) can be performed more efficiently. For each pair of
genotypes, find the rows of allele assignments that allow the pair of
genotypes to share the same allele. This function handles one marker only.
}
\examples{
al.df <- as.data.frame(matrix(c("A","A","A","B","B","A","B","B"), nrow=4))
colnames(al.df) <- c("g1", "g2")
allele_filter(al.df) # list where "g1-g2" maps to (2, 4)

}
