% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compatible_rstrs}
\alias{compatible_rstrs}
\title{Find all vectors of recurrence states compatible with relationship graph}
\usage{
compatible_rstrs(RG, gs_per_ts)
}
\arguments{
\item{RG}{Relationship graph; see \code{\link{enumerate_RGs_alt}}.}

\item{gs_per_ts}{List of vectors of genotypes for each infection.}
}
\value{
Vector of strings (consisting of "C", "L", "I" for recrudescence,
relapse, reinfection respectively) compatible with relationship graph.
}
\description{
Find all vectors of recurrence states compatible with relationship graph
}
\examples{
MOIs <- c(2, 2, 1)
RG <- enumerate_RGs_alt(MOIs, igraph=T)[[175]]
gs_per_ts <- split(paste0("g", 1:sum(MOIs)), rep(1:length(MOIs), MOIs))
# first recurrence cannot be recrudescence, second recurrence cannot be reinfection
plot_RG(RG, edge.curved=0.2)
compatible_rstrs(RG, gs_per_ts) # "LL" "IL" "LC" "IC"

}
