% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pr_IP_RG.R
\name{compute_pr_IP_RG}
\alias{compute_pr_IP_RG}
\title{Compute the probability of an IBD graph given a RG graph}
\usage{
compute_pr_IP_RG(IP, RG)
}
\arguments{
\item{IP}{An IBD graph (IP), which is a partition generated by \code{\link[partition]{listParts}}.}

\item{RG}{A relationship graph (RG), which is an igraph graph; see \code{\link[igraph]{igraph}}.}
}
\value{
A probability from 0 to 1
}
\description{
\code{compute_pr_IP_RG} assumes strangers, siblings, and clones are
represented numerically as 0, 0.5, 1, respectively, and that the probability
of IBD (relatedness) of a given relationship is equal to its numeric
representation, implying that parasite genotypes are sampled from an outbred
population.
}
\examples{

n_genotypes <- 3
RG <- sample_RG(n_genotypes)
IPs <- partitions::listParts(n_genotypes)
compute_pr_IP_RG(IPs[[1]], RG)
plot_RG(RG)

}
