% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{determine_MOIs}
\alias{determine_MOIs}
\title{Determine MOIs from unphased data}
\usage{
determine_MOIs(y)
}
\arguments{
\item{y}{A sequence of lists, where each list contains the genetic data at
each marker as a vector. Each list corresponds to one infection.}
}
\value{
Returns a vector of the (minimum) multiplicity of infection (MOI) for
each infection.
}
\description{
The MOIs returned correspond to the parsimonious explanation of the data,
i.e. the minimal MOIs required to support the observed allelic diversity.
}
\details{
At present, \code{Pv3Rs} only supports prevalence data (categorical data that
signal the detection of alleles), not quantitative data (proportional
abundance) data. The data input expects each per-episode, per-marker allelic
vector to be a set of distinct alleles. Allele repeats at markers with
observed data, and repeat \code{NA}s at markers with missing data, are
removed. \code{NA}s in allelic vectors that also contain non-\code{NA} values
are ignored.
}
\examples{
# two infections, three markers
y <- list(
  list(m1 = c("A", "B"), m2 = c("A"), m3 = c("C")),
  list(m1 = c("B"), m2 = c("B", "C"), m3 = c("A", "B", "C"))
)
determine_MOIs(y) # should be c(2, 3)

}
