% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{enumerate_alleles}
\alias{enumerate_alleles}
\title{Find all allele assignments for genotypes within the same infection}
\usage{
enumerate_alleles(y.inf, gs.inf, use.sym = TRUE)
}
\arguments{
\item{y.inf}{List of alleles observed across markers for genotypes within one
infection. Alleles must not be repeated for the same marker.}

\item{gs.inf}{Vector of genotype names for genotypes within one infection.}

\item{use.sym}{Boolean for permutation symmetry is exploited as a
computational shortcut. Due to permutation symmetry of intra-infection
genotypes, we can fix a single assignment for one of the markers whose
number of alleles observed is equal to the MOI (consider it the anchor)
and permute the rest, discarding combinations that under-represent the
observed marker diversity. The default behaviour is to use this symmetry
such that less assignments have to be considered.}
}
\value{
List of dataframes, one for each marker. The columns correspond to
genotypes, while the rows correspond to allele assignments for a marker.
}
\description{
Note that this function is not tested for input with alleles mixed with NA.
}
\examples{
# 3 markers
y.inf <- list(m1 = c("A", "B"), m2 = c("B", "C", "D"), m3 = c("C"))
enumerate_alleles(y.inf, c("g1", "g2", "g3"))
# 6 assignments for m1 (BAA, ABA, BBA, AAB, BAB, ABB)
# 1 assignment for m2 (accounting for permutation symmetry)
# 1 assignment for m3 (CCC)

}
