% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_RGs.R
\name{enumerate_CPs}
\alias{enumerate_CPs}
\title{Enumerate clonal partitions}
\usage{
enumerate_CPs(MOIs)
}
\arguments{
\item{MOIs}{A numeric vector specifying, for each infection, the number of
distinct parasite genotypes, a.k.a. the multiplicity of infection (MOI).}
}
\value{
A list of all possible partitions, where each partition is encoded
as a membership vector, which indices (genotype names) with the same entry
corresponding to genotypes being int he same partition cell.
}
\description{
A clonal partition is a partition of genotypes where a pair of genotypes of
the same partition cell have a clonal relationship. Genotypes from the same
infection cannot be clones. This code enumerates all clonal partitions,
accounting for this intra-infection restriction; see
vignette("enumerate", package = "Pv3Rs" for more details).
}
\examples{
enumerate_CPs(c(2, 2))

}
