% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_halfsib_alleles.R
\name{enumerate_halfsib_alleles}
\alias{enumerate_halfsib_alleles}
\title{Enumerate all allelic draws for three half siblings}
\usage{
enumerate_halfsib_alleles(n_alleles)
}
\arguments{
\item{n_alleles}{Positive whole number specifying a per-marker number of
alleles, otherwise known as marker cardinality.}
}
\value{
A character matrix. Each column is an individual. Each row is a
possible allelic draw. Alleles are represented by the first \code{n_alleles}
letters of the latin alphabet.
}
\description{
Given a specified number of alleles for a single marker,
\code{generate_halfsib_alleles()} enumerates all the possible ways three half
siblings can draw alleles from their respective parents by firstly
enumerating all allelic combinations for the parents, and by secondly
enumerating all the inheritable combinations for the children.
}
\examples{
generate_halfsib_alleles(3)

}
