% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_IPs.R
\name{enumerate_IPs}
\alias{enumerate_IPs}
\title{Enumerate IBD partitions}
\usage{
enumerate_IPs(MOIs)
}
\arguments{
\item{MOIs}{Numeric vector of multiplicities of infections (MOIs) that gets
reduced to its sum, which must be less than or equal to 11.}
}
\description{
Wrapper function around \code{\link[partitions]{listParts}}. Modifies the
output of \code{\link[partitions]{listParts}} to make genotype names
explicit.
}
\section{To-do}{

Convert to use \code{\link[partitions]{setparts}} whose output takes less
memory (albeit less intuitive), thereby relaxing the limit on \code{MOIs}.
}

\examples{
enumerate_IPs(3)

}
