% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_IPs_RG.R
\name{enumerate_IPs_RG}
\alias{enumerate_IPs_RG}
\title{Enumerate IBD partitions given RG}
\usage{
enumerate_IPs_RG(RG, compat = TRUE)
}
\arguments{
\item{RG}{A relationship graph; see \code{\link{enumerate_RGs}} for details.}

\item{compat}{Logical, if true, the data type of the output matches with the
input IP of \code{\link{compute_pr_IP_RG}}}
}
\description{
The rules are (a) each IBD unit has clonal and sibling relationships only,
(b) clones are IBD, (c) each sibling cluster is part of at most 2 IBD units.
}
\details{
Note that all IBD partitions are equally likely.
}
