% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_RGs.R
\name{enumerate_RGs}
\alias{enumerate_RGs}
\title{Enumerate relationship graphs (RGs)}
\usage{
enumerate_RGs(MOIs, igraph = TRUE, progress.bar = TRUE)
}
\arguments{
\item{MOIs}{Numeric vector of per-episode multiplicities of infection (MOIs),
i.e., number of vertices per episode.}

\item{igraph}{Logical; returns RGs as \code{igraph} objects (default \code{TRUE}).}

\item{progress.bar}{Logical; show progress bar (default \code{TRUE}).}
}
\value{
A list of RGs. If \code{igraph = FALSE},
each RG is a list of length four with:
\describe{
\item{clone}{List of vectors encoding genotypes per clonal cell.}
\item{clone.vec}{Numeric vector with the clonal cell membership of
each genotype.}
\item{sib}{List of vectors encoding clonal cells per sibling cell.}
\item{sib.vec}{Numeric vector with the sibling cell membership of
each clonal cell.}
}
If \code{igraph = TRUE} (default), each RG is encoded as an \code{igraph} object
(see \code{\link{RG_to_igraph}}) encoding equivalent information via edge weights.
}
\description{
An RG is a complete graph over all per-person parasite haploid
genotypes (each as a vertex), with edges labelled clone, sibling, or stranger.
Valid RGs satisfy:
\itemize{
\item{The subgraph induced by the clone edges is a cluster graph.}
\item{The subgraph induced by the clone plus sibling edges is a cluster
graph.}
\item{Clone edges only link genotypes from different episodes.}
}
}
\details{
RGs are generated by enumerating nested set partitions under specific
constraints (see \href{../doc/enumerate.pdf}{vignette("enumerate")} for details).
Clone edges induce a cluster graph, equivalent to a partition of genotypes,
with no clones allowed within the same episode. Sibling edges refine the
clone partition, with no constraints (intra-episode siblings allowed). Each
RG is encoded as a list that encodes the nested set partition. Each partition
is represented in the form of a list of vectors (\code{clone} and \code{sib}) and as a
membership vector (\code{clone.vec} and \code{sib.vec}). By default, an RG encoded as a
list is converted to an \code{igraph} object.
}
\examples{
graphs <- enumerate_RGs(c(2, 1, 2)) # nine graphs

}
