% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_RGs_alt.R
\name{enumerate_RGs_alt}
\alias{enumerate_RGs_alt}
\title{Enumerate transitive relationship graphs, alternate version}
\usage{
enumerate_RGs_alt(MOIs)
}
\description{
This alternate version of \code{enumerate_RGs} is based on generating all
set partitions. This is done twice in a nested fashion: once for determining
clonal relationships, once for determining sibling relationships. A list of
all partitions for sets of difference sizes is pre-computed.
}
