% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{eq_canonical}
\alias{eq_canonical}
\title{Convert equivalence object to canonical form (ie sort lists)}
\usage{
eq_canonical(eq)
}
\description{
Convert equivalence object to canonical form (ie sort lists)
}
