% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lineages.R
\name{generate_lineages}
\alias{generate_lineages}
\title{Generates \code{n} population lineages}
\usage{
generate_lineages(n)
}
\arguments{
\item{n}{Number of lineages to generate.}
}
\description{
Letters represent lineages with repeats if more than 26 lineages are
required.
}
\examples{
rev(generate_lineages(28))

}
