% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hash.IP}
\alias{hash.IP}
\title{Convert IBD partition to a unique string for hashing}
\usage{
hash.IP(IP, gs)
}
\arguments{
\item{IP}{List containing vectors of genotype names corresponding to IBD
clusters.}

\item{gs}{Vector containing all genotype names.}
}
\value{
Membership vector in string format.
}
\description{
This is used for building a hash table for p(y at marker m|IBD). Works up to
10 genotypes. This seems to be marginally faster than
\code{\link{convert_IP_to_string}}.
}
