% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.clone.compatible.R
\name{is.clone.compatible}
\alias{is.clone.compatible}
\title{Internal check function}
\usage{
is.clone.compatible(A, vertex_data_matrix)
}
\description{
Check if Pr_yn_G > 0 due to a clonal edge between non-identical genotypes
}
\section{Provenance}{
 This function was adapted from
\code{test_cln_incompatible} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.

@noRd
}

