% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Function to test if x contains integer numbers.}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\description{
This function is copied from \code{\link[base]{integer}}. It is differs from \code{\link[base]{is.integer}}
because tests if x is an integer type, not an integer number.
}
