% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages_to_IP.R
\name{lineages_to_IP}
\alias{lineages_to_IP}
\title{Converts a vector of per-genotype lineages to an IBD partition.}
\usage{
lineages_to_IP(lin_vec)
}
\arguments{
\item{lin_vec}{A character vector of parasites lineages whose names are genotype
indices.}
}
\description{
\code{lineage_to_IP} is copied from \code{\link[partitions]{vec_to_eq}} and
modified so that the output is comparable with the output of
\code{\link[partitions]{listParts}}. Otherwise stated, the output is ordered so that large
clusters precede small clusters and so that low genotype indices precede
high genotype indices.
}
\examples{
RG <- sample_RG(3)
lineages <- sample_lineages_from_infinite_pop(RG)
lineages_to_IP(lineages)

}
