% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_type_summary.R
\name{locus_type_summary}
\alias{locus_type_summary}
\title{Summarise locus-wise data types}
\usage{
locus_type_summary(y)
}
\arguments{
\item{y}{A list of lists for two episodes; see \code{\link[=compute_posterior]{compute_posterior()}} for more
details.}

\item{m}{A postive whole number indexing a marker.}
}
\value{
A vector of strings summarising the data at each locus.
}
\description{
Summarise locus-wise data types
}
\details{
A function to summarise the data at each locus as one of four types:
\itemize{
\item All match (all genotypes have the same allele).
\item All diff. (all genotypes have a different allele).
\item Intra-match (some intra-episode genotypes have the same allele)
\item Intra-match (some inter-episode genotypes have the same allele).

The number of apparent genotypes is the sum of the per-episode MOIs. When
the total number of apparent genotypes exceeds 3, "Intra-match" excludes
any "Inter-match" (i.e., it is equivalent to Inter-matches only), whereas
an "Inter-match does not exclude an "Intra-match
}
}
\examples{
# example code
y <- list(
  list(m1 = c("A", "B"), m2 = c("A"), m3 = c("C")),
  list(m1 = c("B"), m2 = c("B", "C"), m3 = c("A", "B", "C", "D"))
)

locus_type_summary
}
