% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Pr_yn_Gab.R
\name{Log_Pr_yn_Gab}
\alias{Log_Pr_yn_Gab}
\title{Internal computation function featuring in Log_Pr_yn_Gnb_unnormalised}
\usage{
Log_Pr_yn_Gab(G, log_Fs, MSs, Z, A, vertex_data_matrix)
}
\description{
Notes copied from original:
MS where one or more NA are ignored when summing over m
G is assumed labelled
Z is a vector of alpha_terms (calculated above since not data dependent
}
\section{Provenance}{
 This function was adapted from
\code{Log_Pr_yn_Gab} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.
}

