% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Pr_yn_Gab_unnormalised.R
\name{Log_Pr_yn_Gnb_unnormalised}
\alias{Log_Pr_yn_Gnb_unnormalised}
\title{Internal computation function featuring in post_prob_CLI()}
\usage{
Log_Pr_yn_Gnb_unnormalised(G, Gabs, cn, Tn, log_Fs, MSs, alpha_terms)
}
\description{
Notes copied from original:
Function to sum over vertex haploid genotype labels given edge kinship
labelled graph Returns: Pr(yn | Gnb) unnormalised = sum from a = 1 to A over
Pr(yn | Gnab) Unnormalised because  Pr(yn | Gnb) = sum from a = 1 to A over
Pr(yn | Gnab) * (1/A)
}
\section{Provenance}{
 This function was adapted from \code{Log_Pr_yn_Gnb_unnormalised} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.
}

