% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maxima}
\alias{maxima}
\title{Upper bounds on posterior probabilities}
\format{
A matrix with 4 rows and 247 columns:
\describe{
\item{C_with}{Upper bound on the posterior probability of recrudescence
when summation includes graphs with cliques of three or more intra-episode siblings}
\item{C_wout}{Upper bound on the posterior probability of recrudescence
when summation excludes graphs with cliques of three or more intra-episode siblings}
\item{I_with}{Upper bound on the posterior probability of reinfection when
summation includes graphs with cliques of three or more intra-episode siblings}
\item{I_wout}{Upper bound on the posterior probability of reinfection when
summation excludes graphs with cliques of three or more intra-episode siblings}
\item{11}{MOI vector (1, 1)}
\item{12}{MOI vector (1, 2)}
...
\item{11111111}{MOI vector (1, 1, 1, 1, 1, 1, 1, 1)}
}
}
\source{
\url{https://github.com/aimeertaylor/Pv3Rs/blob/main/data-raw/maxima.R}
}
\usage{
maxima
}
\description{
Upper bounds on posterior probabilities of recrudescence and reinfection
(rows) for all MOI vectors that sum to a total genotype count of eight
(columns). Upper bounds are induced by the prior \emph{plus} knowledge that all but
the first recurrence has data when the recurrence count exceeds one. We
report two upper bounds for each recurrent state: one where summation over
graphs includes graphs with cliques of three or more intra-episode siblings;
another where summation over graphs excludes graphs with cliques of three or
more intra-episode siblings. Since graphs with cliques of three or more
intra-episode siblings only apply when MOI vectors include MOIs of three or
more, these two bounds only differ for MOI vectors that include MOIs of three
or more (see example below). The column names represent MOI vectors whose
MOIs are all less than or equal to seven.
}
\examples{
# Convert column names to MOI character vectors
MOIvec <- strsplit(colnames(maxima), split = "")

# Get MOI vectors with MOIs that exceed two
Exceed2 <- sapply(MOIvec, function(x) any(as.numeric(x) > 2))

# Compare reinfection upper bounds with and with summation for MOI vectors that exclude
# and include MOIs greater than two
all(maxima["I_with", !Exceed2] == maxima["I_wout", !Exceed2])
any(maxima["I_with", Exceed2] == maxima["I_wout", Exceed2])
}
\keyword{datasets}
