% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{class}
\name{msg_progress_bar-class}
\alias{msg_progress_bar-class}
\alias{msg_progress_bar}
\title{Message progress bar}
\description{
A simple progress bar to use in R packages where messages are preferred to
console output. See https://gist.github.com/MansMeg/1ec56b54e1d9d238b4fd.
}
\section{Fields}{

\describe{
\item{\code{iter}}{Total number of iterations}

\item{\code{i}}{Current iteration}

\item{\code{width}}{Width of the R console}

\item{\code{width_bar}}{Width of the progress bar}

\item{\code{progress}}{The number of character printed (continous)}

\item{\code{progress_step}}{Addition to progress per iteration}
}}

\section{Methods}{

\describe{
\item{\code{increment()}}{A messagebar object.}

\item{\code{initialize(iter)}}{Initialize a messagebar object}
}}

\author{
Mans Magnusson (MansMeg @ github)
}
