% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plots the data}
\usage{
plot_data(
  ys,
  fs = NULL,
  patient.vert = FALSE,
  mar = c(1.5, 3.5, 1.5, 1),
  gridlines = TRUE,
  palette = RColorBrewer::brewer.pal(12, "Paired"),
  marker.annotate = TRUE,
  legend.lab = "Allele frequencies",
  legend.line = 0.2,
  legend.ylim = c(0.05, 0.2),
  cex.maj = 0.7,
  cex.min = 0.5,
  cex.text = 0.5,
  x.line = 0.2,
  y.line = 2.5
)
}
\arguments{
\item{ys}{A nested list of per-patient, per-episode, per-marker allelic data.
Specifically, a per-patient list of a per-episode list of a per-marker list
of character vectors of observed alleles.}

\item{fs}{A per-marker list of numeric vectors of allele frequencies. If NULL
(default), for a given marker, only the alleles present in the data are
represented in the legend, and each allele is represented equally. Because
the colour scheme is adaptive (see introduction), the same allele will have
a different colour in a plot of an alternative data list if more or fewer
alleles are observed at the given marker across the alternative data list.
If fs is specified, all possible alleles are represented and legend areas
are proportional to allele frequencies; i.e., common alleles have relatively
large legend areas, and rare alleles have relatively small legend areas.
Specify fs to fix the colour of a given allele across plots of different
data lists, thereby facilitating cross-comparison.}

\item{patient.vert}{Logical. If true, patient IDs are printed vertically;
otherwise they are printed horizontally (default).}

\item{mar}{A vector which gives the number of lines of margin for the main
plot; see the entry for \code{mar} in \code{\link[graphics]{par}}.}

\item{gridlines}{Logical. If true (default), white gridlines separating
patients and markers will be drawn.}

\item{palette}{Colour palette for alleles, see return value of
\code{\link[RColorBrewer]{brewer.pal}}. Colour interpolation takes place
when the number of alleles exceeds the number of colours defining the
palette.}

\item{marker.annotate}{Logical. If true (default), the names of the alleles
are printed on top of their colours in the legend.}

\item{legend.lab}{Label for the axis of the color legend, defaults to "Allele frequencies".
Set to \code{NA} if not desired. If this is the case, consider adjusting
\code{legend.ylim} to use more plotting space.}

\item{legend.line}{Distance in units of character height of the legend label
from the colour bar, defaults to 1.5.}

\item{legend.ylim}{Vector of two floats representing the y-coordinate limits
of the legend in device coordinates (between 0 and 1). Defaults to c(0.05, 0.2).}

\item{cex.maj}{Float representing the font scaling of major axis labels.}

\item{cex.min}{Float representing the font scaling of minor axis labels.}

\item{cex.text}{Float representing the font scaling for the allele labels.}

\item{x.line}{Distance between top x-axis and x-axis label, defaults to 0.2.}

\item{y.line}{Distance between left y-axis and y-axis label, defaults to 2.5.}
}
\description{
Plots the alleles (colours), which are observed in different episodes (rows),
on different markers (columns), where episodes are grouped by patient. The
patients and per-patient episodes are plotted from bottom to top. If more
than one allele is detected per episode per marker, the corresponding
row-column entry is subdivided into different colours. The order of markers
follow the allele frequencies, if provided. Otherwise, markers are ordered
lexicographically.
}
\details{
The legend depicts the alleles of the markers as the markers appear from left
to right in the main plot. Otherwise stated, the legend is ordered by the
order of markers stated on on the vertical axis of the main plot. The default
colour scheme is adaptive. It is designed to visually differentiate the
alleles as much as  ossible: the maximum range of qualitative scheme, with
contrast of hue between adjacent colours, is always used; the adjacent
colours areinterpolated only if a given marker has more than 12 alleles. The
names of the alleles are printed on top of their colours if
\code{marker.annotate} is set to \code{TRUE}.
}
\examples{

# Plot example Plasmodium vivax data set
mar <- c(2, 3.5, 1.5, 1) # extra vertical margin for vertical patient labels
plot_data(ys = ys_VHX_BPD, patient.vert = TRUE, mar = mar, legend.lab = NA)
plot_data(ys = ys_VHX_BPD, fs = fs_VHX_BPD, patient.vert = TRUE, mar = mar,
          legend.lab = NA)
plot_data(ys = ys_VHX_BPD, fs = fs_VHX_BPD, marker.annotate = FALSE,
          patient.vert = TRUE, mar = mar, legend.lab = NA)

# Demonstrating the adaptive nature of the colour scheme:
ys <- ys_VHX_BPD["VHX_52"] # A single patient
plot_data(ys, fs = fs_VHX_BPD) # Colours and the legend match plots above
plot_data(ys) # Colours and the legend adapt to alleles detected in VHX_52


}
