% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RG.R
\name{plot_RG}
\alias{plot_RG}
\title{Plot a relationship graph (RG)}
\usage{
plot_RG(
  RG,
  edge_col = c(`0` = "white", `0.5` = "black", `1` = "black"),
  edge_lty = c(`0` = "solid", `0.5` = "dashed", `1` = "solid"),
  vertex_palette = "Set2",
  layout_by_group = TRUE,
  edge.width = 1.5,
  ...
)
}
\arguments{
\item{RG}{A relationship graph, which is an igraph graph; see
\code{\link{enumerate_RGs}}.}

\item{edge_col}{A vector of edge colours corresponding to different
relationships, where 0 represents a stranger, 0.5 represents a sibling and
1 represents a clone.}

\item{edge_lty}{A vector of edge line types corresponding to different
relationships, where 0 represents a stranger, 0.5 represents a sibling and
1 represents a clone.}

\item{vertex_palette}{A character string specifying a RColorBrewer palette.
Overrides the default \code{palette} of \code{\link[igraph]{plot.igraph}}.}

\item{layout_by_group}{A logical argument which if TRUE (default) overrides
the default layout of \code{\link[igraph]{plot.igraph}} so that vertices
that represent parasite genotypes from different infections are distributed
horizontally and vertices that represent genotypes within infections are
distributed vertically.}

\item{edge.width}{Overrides the default \code{edge.width} of
\code{\link[igraph]{plot.igraph}}.}

\item{...}{Additional arguments to pass to \code{\link[igraph]{plot.igraph}}, e.g.
\code{edge.curved}.}
}
\description{
This function is a wrapper around \code{\link[igraph]{plot.igraph}}, designed to group
genotypes by infections, both spatially and using vertex colour. It also makes sure
clonal and sibling edges are plotted differently using different line types.
}
\section{Provenance}{
 This function was adapted from \code{plot_Vivax_model} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.
}

\examples{
RGs <- enumerate_RGs(c(1,2,1))
for(i in 1:length(RGs)) plot_RG(RGs[[i]], edge.curved = 0.5)

}
