% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_data}
\alias{prep_data}
\title{Pre-process data to remove repeats and \code{NA}s}
\usage{
prep_data(y)
}
\arguments{
\item{y}{A sequence of lists, where each list contains the genetic data at
each marker as a vector. Each list corresponds to one infection.}
}
\description{
Removes repeated alleles and any \code{NA}s in allelic vectors that also
contain non-\code{NA} values.
}
