% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pv3Rs.R
\docType{package}
\name{Pv3Rs}
\alias{Pv3Rs}
\title{Pv3Rs: A package for inferring \emph{Plasmodium vivax} (Pv) recrudescence, reinfection, and relapse (3Rs).}
\description{
This description is not yet finished; see 10.6 of https://r-pkgs.org/man.html#text-formatting.
}
\details{
Consider naming the package differently to encourage Pf2Rs
(or importing Pv3Rs into a different package called Pf2Rs).

The Pv3Rs package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

