% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine_parent_ids.R
\name{recombine_parent_ids}
\alias{recombine_parent_ids}
\title{Generate parental allocations for a meiotic tetrad}
\usage{
recombine_parent_ids(per_chr_marker_counts)
}
\arguments{
\item{per_chr_marker_counts}{A numeric vector of marker counts per chromosome}
}
\description{
For a given set of per chromosome marker counts, generate per marker parental
ids for a meiotic tetrad generated by sexual recombination (chromosomal
crossovers followed by independent assortment). We assume
\enumerate{
\item For all chromosomes, both pairs of non-sister chromatids cross over
\item One crossover per non-sister chromatid pair
\item Equal-length chromosomes (follows from above; in reality, chromosomes
have different lengths, and the number of crossovers increases with length)
\item Equi-distributed markers (implicit)
}
}
\examples{
n_chrs <- 14 # P. vivax has 14 chromosomes
n_markers <- 100 # For 100 markers
chr_per_marker <- round(seq(0.51, n_chrs + 0.5, length.out = n_markers))
n_markers_per_chr <- table(chrs_per_marker) # Per chromosome marker counts
recombine_parent_ids(n_markers_per_chr)

}
