% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG_inference.R
\name{RG_inference}
\alias{RG_inference}
\title{Computes the likelihood of all relationship graphs}
\usage{
RG_inference(MOIs, fs, alleles_per_m)
}
\arguments{
\item{MOIs}{A numeric vector specifying, for each infection, the number of
distinct parasite genotypes, a.k.a. the multiplicity of infection (MOI).}

\item{fs}{List of frequencies of each allele at each marker.}

\item{alleles_per_m}{List of allele assignments for each marker (see the
\code{al.mat} parameter to \code{\link{allele_filter}}).}
}
\value{
List of all relationship graphs, including their log-likelihoods as
a variable under each relationship graph.
}
\description{
Computes the likelihood of all relationship graphs
}
\examples{
# 1 marker, 2 infections, MOIs = 2, 1
MOIs <- c(2, 1)
fs <- list(m1=setNames(c(0.4, 0.6), c("A", "B")),
           m2=setNames(c(0.2, 0.8), c("C", "D")))
al_df1 <- as.data.frame(matrix(c("A","B","B"), nrow=1))
al_df2 <- as.data.frame(matrix(c("C","D","C",
                                  "D","C","C"), nrow=2, byrow=T))
colnames(al_df1) <- colnames(al_df2) <- paste0("g", 1:3)
alleles_per_m <- list(m1=al_df1, m2=al_df2)
RGs <- RG_inference(MOIs, fs, alleles_per_m)

}
