% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_IP_given_RG.R
\name{sample_IP_given_RG}
\alias{sample_IP_given_RG}
\title{Sample an IBD partition given a relationship graph}
\usage{
sample_IP_given_RG(RG, outbred = TRUE, ...)
}
\arguments{
\item{RG}{A relationship graph; see \code{\link{enumerate_RG}}.}

\item{outbred}{Logical specifying whether or not the parasite genotypes belong to
an outbred population or not.}

\item{...}{Additional arguments to be passed to
\code{\link{sample_lineages_from_finite_pop}} if \code{outbred = FALSE}.}
}
\description{
\code{sample_IP_given_RG} is a wrapper function that converts lineages
sampled by \code{\link{sample_lineages_from_infinite_pop}} (default) or by
\code{\link{sample_lineages_from_finite_pop}} (if \code{outbred = FALSE})
into an IBD partition; see enumerate_IP.
}
\examples{
RG <- sample_RG(3)
sample_IP_given_RG(RG)

}
