% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_IPs_given_RG.R
\name{sample_IPs_given_RG}
\alias{sample_IPs_given_RG}
\title{Sample IBD partition/s given a relationship graph}
\usage{
sample_IPs_given_RG(RG, n_m, outbred = TRUE, ...)
}
\arguments{
\item{RG}{A relationship graph; see \code{\link{enumerate_RG}}.}

\item{n_m}{Number of markers.}

\item{outbred}{Logical specifying whether or not the parasite genotypes belong to
an outbred population or not.}

\item{...}{Additional arguments to be passed to
\code{\link{sample_lineages}} when \code{outbred = FALSE}.}
}
\description{
\code{sample_IPs_given_RG} is a wrapper function that converts per-marker lineages
sampled by \code{\link{sample_lineages}} into per-marker IBD partitions; see enumerate_IP.
}
\examples{
set.seed(1)
RG <- sample_RG(3)
sample_IPs_given_RG(RG, n_m = 3)

}
