% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_RG.R
\name{sample_RG}
\alias{sample_RG}
\title{Sample a relationship graph (RG)}
\usage{
sample_RG(MOIs, igraph = TRUE)
}
\arguments{
\item{MOIs}{Numeric vector containing per-episode multiplicities of infection (MOIs).}

\item{igraph}{Logical; returns RG as an \code{igraph} object (default \code{TRUE}).}
}
\value{
An RG; see \code{\link{enumerate_RGs}}) for more details.
}
\description{
Uses the techniques in \code{\link{enumerate_RGs}} to sample a relationship
graph uniformly at random. All clonal partitions are generated, and the
number of sibling partitions consistent with each clonal partition is
determined. A clonal partition is randomly selected with probability
proportional to the corresponding number of sibling partitions, and a sibling
partition is then uniformly sampled. The nested partition is equivalent to a
relationship graph. See \code{\link{enumerate_RGs}} for details on the nested
partition representation of a relationship graph.
}
\examples{
set.seed(1)
RG <- sample_RG(c(3, 2))
plot_RG(RG)

}
