% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_RG_alt.R
\name{sample_RG_alt}
\alias{sample_RG_alt}
\title{Sample a transitive relationship graph, alternate version}
\usage{
sample_RG_alt(MOIs, igraph = T)
}
\arguments{
\item{MOIs}{A numeric vector specifying, for each infection, the number of
distinct parasite genotypes, a.k.a. the multiplicity of infection (MOI).}

\item{igraph}{Logical for whether to return an \code{igraph} object.}
}
\value{
See \code{\link{enumerate_RGs_alt}}.
}
\description{
Uses the techniques in \code{\link{enumerate_RGs_alt}} to uniformly sample
a relationship graph.
}
