% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_single_recurrence_RG.R
\name{sample_single_recurrence_RG}
\alias{sample_single_recurrence_RG}
\title{Sample a transitive relationship graph (RG) for a single recurrence}
\usage{
sample_single_recurrence_RG(
  MOI_init,
  cause,
  prob_recrudesce = 0.5,
  beta_relapse = 1
)
}
\arguments{
\item{MOI_init}{The number of distinct parasite genotypes (a.k.a the
multiplicity of infection, MOI) of the initial infection.}

\item{cause}{A character string specifying the cause of the recurrent
infection: \code{"recrudescence"}, \code{"reinfection"} or
\code{"relapse"}.}

\item{prob_recrudesce}{A probability that each genotype survives in a
recrudescent infection.}

\item{beta_relapse}{A multiplicative factor that modifies the average MOI of
a reinfection to obtain the average MOI of a relapse.}
}
\value{
Returns a transitive relationship graph (RG); see \code{\link{enumerate_RGs}} for details.
}
\description{
Samples one transitive graph of stranger and sibling relationships between
distinct parasite genotypes within the initial and recurrent infections and,
depending on the specified cause, stranger, sibling and/or clonal
relationships across parasite genotypes in the initial and recurrent
infections. The function requires the user to specify the multiplicity of the
initial infection, \code{MOI_init}. If \code{MOI_init} is greater than 4,
\code{sample_single_recurrence_RG} is liable to take a long time because it is
testing for transitivity while sampling among not necessarily transitive
graphs, of which there will be many; see the message that
\code{sample_single_recurrence_RG} prints to the screen.
}
\section{To-do}{

Better to test transitivity on intra before whole?
}

\examples{

for (cause in c("recrudescence", "reinfection", "relapse")) {
  RG <- sample_single_recurrence_RG(MOI_init = 3, cause)
  plot_RG(RG)
  print(RG)
}


}
