% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_two}
\alias{split_two}
\title{Partition a vector into at most two subvectors}
\usage{
split_two(s)
}
\arguments{
\item{s}{Vector to be split.}
}
\value{
Given a vector with no repeats, returns a list consisting of
\itemize{
\item a list that contains the original vector as its only element
\item other lists where each list contains two disjoint vectors whose
union covers the vector. All possible unordered pairs are included.
}
}
\description{
Partition a vector into at most two subvectors
}
\examples{
gs <- paste0("g", 1:3)
# 4 possibilities in total
# either all genotypes in one vector (1 possibility)
# or 2 genotypes in one vector and the last in one vector (3 possibilties)
split_two(gs)

}
