% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cln_compatible.R
\name{test_cln_compatible}
\alias{test_cln_compatible}
\title{Internal check function}
\usage{
test_cln_compatible(A, vertex_data_matrix)
}
\description{
Check if Pr_yn_G > 0 due to a clonal edge between non-identical genotypes
}
\section{Provenance}{
 This function was adapted from
\code{test_cln_incompatible} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.
}

