% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_transitive.R
\name{test_transitive}
\alias{test_transitive}
\title{A function to test if a relationship graph is transitive.}
\usage{
test_transitive(RG)
}
\arguments{
\item{RG}{A relationship graph; see \code{\link{enumerate_RGs}}.}
}
\description{
\code{test_transitive} implements the brute-force algorithm described in the
appendix of Taylor & Watson et al. Nature communications (2019). It assumes
there are only three relationships: strangers, siblings and clones,
represented by weights 0, 0.5 and 1, respectively, such that strangers are
disconnected, whereas siblings and clones are connected. Consequently, all
components of a transitive relationship graph must be cliques. Meanwhile,
among cliques of size three, an edge weight sum of 2.5 is indicative of a
intransitive clone + clone + sibling trio.
}
\section{Provenance}{
 This function was adapted from
test_correct_graph at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}
}

