% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Computes the posterior distribution of recurrence states}
\usage{
compute_posterior(y, fs, prior = NULL, return.RG = FALSE)
}
\arguments{
\item{y}{Observed data in the form as a list. The number of entries is the
number of episode. Each entry is in turn a list of observed alleles
for each marker.}

\item{fs}{List of allele frequencies. Names of the list must match with the
marker names in \code{y}.}

\item{prior}{Matrix of prior probabilities of the recurrence states for each
recurrent episode. Each row corresponds to an episode, and the column names
must be C, L, and I. If prior not provided, a uniform prior will be used.}

\item{return.RG}{Boolean for whether to return the relationship graphs, defaults
to \code{FALSE}.}
}
\value{
List containing:
\describe{
\item{marg}{Matrix of marginal posterior probabilities of the possible
recurrence states for each recurrent episode.}
\item{joint}{Vector of joint posterior probabilities of each possible string
of recurrence states.}
\item{RGs}{List of relationship graphs with their log-likelihoods stored.
Only returned if \code{return.RG} is \code{TRUE}.}
}
}
\description{
Computes the posterior distribution of recurrence states
}
\details{
Assumptions:
\itemize{
\item{No genotyping errors and missing data}
\item{Parasites are outbred}
\item{Relationship graphs are equally likely}
\item{Uniform prior over recurrent states}
}
}
\examples{
y <- list(list(m1=c("A","B"), m2=c("C")), list(m1=c("A"), m2=c("C","D")))
fs <- list(m1=setNames(c(0.4, 0.6), c("A","B")),
           m2=setNames(c(0.2, 0.8), c("C","D")))
prior <- matrix(c(0.2, 0.3, 0.5), nrow=1)
post <- compute_posterior(y, fs, prior)
post$marg

}
