% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pr_partitions.R
\name{compute_pr_partitions}
\alias{compute_pr_partitions}
\title{Compute the probabilities of partitions}
\usage{
compute_pr_partitions(k, f)
}
\arguments{
\item{k}{Number of balls to partition}

\item{f}{Probability that two balls are in the same box}
}
\description{
Computes the probabilities of all possible partitions of k balls given each
pair of balls has probability f of being in the same box (i.e. there are 1/f
equally probable boxes). Can be used to compute the probability of IBD
partitions among parasites who are all equally related to one another with
relatedness f. Can also be used to inefficiently solve the birthday problem
provided that there are not too many balls.
}
\examples{

compute_pr_partition(3, f = 0.5) # Agrees with IBD computations by hand
compute_pr_partition(3, f = 1) # Agrees with IBD computations by hand
compute_pr_partition(3, f = 0) # Agrees with IBD computations by hand

# Using compute_pr_partition to solve the birthday problem with...
k <- 5 # people
days <- 365 # days in the year

# birthday problem:
stats::pbirthday(n = k, classes = days, coincident = 2)

# birthday problem computed using compute_pr_partition:
pr_parts <- compute_pr_partition(k, f = 1/days)
1-tail(pr_parts,1)
}
