% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_IP_to_string.R
\name{convert_IP_to_string}
\alias{convert_IP_to_string}
\title{Converts a IBD partition into a character string}
\usage{
convert_IP_to_string(IP)
}
\arguments{
\item{IP}{IBD partition}
}
\description{
An IBD partition is an equivalence list that can be converted into a string.
}
\examples{
IPs <- enumerate_IPs(3); IPs
sapply(IPs, convert_IP_to_string)

}
