% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{determine_MOIs}
\alias{determine_MOIs}
\title{Determine MOIs from unphased data}
\usage{
determine_MOIs(y)
}
\arguments{
\item{y}{A sequence of lists, where each list contains the genetic data at
each marker for one infection.}
}
\value{
Returns a vector of the (minimum) multiplicity of infection (MOI) for
each infection.
}
\description{
Determine MOIs from unphased data
}
\examples{
# two infections, three markers
y <- list(list(m1=c("A","B"), m2=c("A"), m3=c("C")),
          list(m1=c("B"), m2=c("B","C"), m3=c("A","B","C")))
determine_MOIs(y) # should be c(2, 3)

}
