% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_MOIs.R
\name{determine_MOIs}
\alias{determine_MOIs}
\title{Determine multiplicities of infection (MOIs)}
\usage{
determine_MOIs(y, return.names = FALSE)
}
\arguments{
\item{y}{List of lists encoding allelic data; see
\code{\link{compute_posterior}} for more details. The outer list contains
episodes in chronological order. The inner list contains named markers per
episode. For each marker, one must specify an allelic vector: a set of
distinct alleles detected at that marker; or \code{NA} if marker data are
missing.}

\item{return.names}{Logical; if TRUE and \code{y} has named episodes, episode
names are returned.}
}
\value{
Numeric vector containing one MOI per episode, each MOI representing
the maximum number of alleles observed at any marker per episode.
}
\description{
Calculates a MOI for each episode based on allelic diversity across markers.
}
\details{
The MOI of an episode is the number of genetically distinct groups of clonal
parasites within the episode. All parasites within a group share the same DNA
sequence, which we call a genotype. As such, a MOI is also a distinct
parasite genotype count.

For a given episode, the MOI is the maximum number distinct alleles observed
at any marker within that episode. These MOIs are the default used by
\code{\link{compute_posterior}}.
}
\examples{
y <- list(enrol = list(m1 = c("A", "B"), m2 = c("A"), m3 = c("C")),
          recur = list(m1 = c("B"), m2 = c("B", "C"), m3 = c("A", "B", "C")))
determine_MOIs(y) # returns c(2, 3)

}
