% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{enumerate_alleles}
\alias{enumerate_alleles}
\title{Find all allele assignments for genotypes within the same infection}
\usage{
enumerate_alleles(y.inf, gs.inf)
}
\arguments{
\item{y.inf}{List of alleles observed across markers for genotypes within
one infection.}

\item{gs.inf}{Vector of genotype names for genotypes within one infection.}
}
\value{
List of dataframes, one for each marker. The columns correspond to
genotypes, while the rows correspond to allele assignments for a marker.
}
\description{
Due to permutation symmetry of intra-infection genotypes, we fix a single
assignment for one of the markers whose number of alleles observed is equal
to the MOI (consider it the anchor) and permute the rest, discarding
combinations that under-represent the observed marker diversity.
}
\examples{
# 3 markers
y.inf <- list(m1 = c("A", "B"), m2 = c("B", "C", "D"), m3 = c("C"))
enumerate_alleles(y.inf, c("g1", "g2", "g3"))
# 6 assignments for m1 (BAA, ABA, BBA, AAB, BAB, ABB)
# 1 assignment for m2 (accounting for permutation symmetry)
# 1 assignment for m3 (CCC)

}
