% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_IPs_RG.R
\name{enumerate_IPs_RG}
\alias{enumerate_IPs_RG}
\title{Enumerate IBD partitions consistent with a given relationship graph; see
vignette("enumerate", package = "Pv3Rs") for more details.}
\usage{
enumerate_IPs_RG(RG, compat = TRUE)
}
\arguments{
\item{RG}{A relationship graph; see \code{\link{enumerate_RGs}} for details.}

\item{compat}{Logical, if true, a list of \code{partitions} equivalence
objects are returned. Otherwise, returns a list of IBD parititon vectors in
reference to the sibling clusters.}
}
\description{
Enumerate IBD partitions consistent with a given relationship graph; see
vignette("enumerate", package = "Pv3Rs") for more details.
}
\details{
For a IBD partition to be consistent with the relationship graph given, it
must satisfy the following:
\itemize{
\item{genotypes within each IBD cell have clonal and sibling relationships
only,}
\item{genotypes that are clones must be in the same IBD cell,}
\item{each cluster of sibling units span at most 2 IBD cells (corresponds
to two parents).}
}

Note that all IBD partitions are equally likely.
}
\examples{
set.seed(3)
RG <- sample_RG(c(2, 1, 2))
enumerate_IPs_RG(RG)

}
