% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hash.IP}
\alias{hash.IP}
\title{Convert IBD partition to a unique string for hashing}
\usage{
hash.IP(IP, gs)
}
\arguments{
\item{IP}{List containing vectors of genotype names with each vector
corresponding to an IBD cell.}

\item{gs}{Vector containing all genotype names.}
}
\value{
String where the integers in the IBD membership vector have been
converted to ASCII characters.
}
\description{
This is used for building a hash table for p(y at marker m|IBD).
}
\examples{
gs <- paste0("g", 1:3)
IP1 <- list(c("g1", "g3"), c("g2"))
IP2 <- list(c("g2"), c("g3", "g1"))
hash1 <- hash.IP(IP1, gs)
hash2 <- hash.IP(IP2, gs)
hash1 == hash2 # TRUE, even though the order is different

}
