% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RG.R
\name{plot_RG}
\alias{plot_RG}
\title{Plot a relationship graph (RG)}
\usage{
plot_RG(
  RG,
  layout_by_group = TRUE,
  vertex_palette = "Set2",
  edge_lty = c(`0.5` = "dashed", `1` = "solid"),
  edge_col = c(`0.5` = "black", `1` = "black"),
  edge_width = 1.5,
  ...
)
}
\arguments{
\item{RG}{A relationship graph, which is an \code{igraph} graph; see return
value of \code{\link{RG_to_igraph}}.}

\item{layout_by_group}{A logical argument which if TRUE (default) overrides
the default layout of \code{\link[igraph]{plot.igraph}} so that vertices
that represent parasite genotypes from different infections are distributed
horizontally and vertices that represent genotypes within infections are
distributed vertically.}

\item{vertex_palette}{A character string specifying an RColorBrewer palette.
Overrides the default \code{palette} of \code{\link[igraph]{plot.igraph}}.}

\item{edge_lty}{A vector of edge line types corresponding to different
relationships, where 0.5 represents a sibling and 1 represents a clone.}

\item{edge_col}{A vector of edge colours corresponding to different
relationships, where 0.5 represents a sibling and 1 represents a clone.}

\item{edge_width}{Overrides the default \code{edge.width} of
\code{\link[igraph]{plot.igraph}}.}

\item{...}{Additional arguments to pass to \code{\link[igraph]{plot.igraph}}, e.g.
\code{edge.curved}.}
}
\description{
This function is a wrapper around \code{\link[igraph]{plot.igraph}}, written
to group parasite genotypes by infection, both spatially and using vertex
colour. Specifically, parasite genotypes within infections are vertically
distributed with some horizontal jitter when \code{layout_by_group} is TRUE
(default), and coloured the same. It also makes sure clonal and sibling edges
are plotted differently using different line types.
}
\section{Provenance}{
 This function was adapted from \code{plot_Vivax_model} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.
}

\examples{
RGs <- enumerate_RGs(c(2, 1, 1))
cpar <- par() # record current par before changing
par(mfrow = c(3, 4), mar = c(0.1, 0.1, 0.1, 0.1))
for (i in 12:23) {
  plot_RG(RGs[[i]], edge.curved = 0.1)
  box()
}
par(cpar) # reset par
}
