% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simplex.R
\name{plot_simplex}
\alias{plot_simplex}
\title{Plots a 2D simplex}
\usage{
plot_simplex(
  v.labels = c("Recrudescence", "Relapse", "Reinfection"),
  v.cutoff = 0.5,
  v.colours = c("yellow", "purple", "red"),
  plot.tri = TRUE,
  p.coords = NULL,
  p.labels = rownames(p.coords),
  p.labels.pos = 3,
  ...
)
}
\arguments{
\item{v.labels}{Vector of labels that annotate vertices anticlockwise from
top (default: "Recrudescence", "Relapse", "Reinfection"). If NULL, vertices
are not annotated.}

\item{v.cutoff}{Number between 0.5 and 1 that separates regions
of lower and higher probability. Beware the use of cut-offs for probable
recrudescence classification and probable reinfection classification; see
\href{https://aimeertaylor.github.io/Pv3Rs/articles/understand-posterior.html}{"Understand posterior estimates"}.}

\item{v.colours}{Vector of colours associated with the vertices
anticlockwise from top.}

\item{plot.tri}{Logical; plots the triangular boundary if \code{TRUE} (default).}

\item{p.coords}{Matrix of 3D simplex coordinates (e.g., per-recurrence
probabilities of recrudescence, relapse and reinfection), one set of
coordinates per row, each row is projected onto 2D coordinates using
\code{\link[=project2D]{project2D()}}, then plotted as a single simplex point using
\code{\link[graphics:points]{graphics::points()}}. If the user provides a vector, it is converted to a
matrix with one row.}

\item{p.labels}{Labels used to annotate points given by \code{p.coords}.
Equal to the row names of \code{p.coords} by default. Points are not
annotated if \code{NA}.}

\item{p.labels.pos}{Position specifier for \code{p.labels}. Values of \code{1},
\code{2}, \code{3} and \code{4}, respectively indicate positions below, to
the left of, above and to the right of the points. Can be either a single
integer (default 3) or a vector of integers.}

\item{...}{Further graphical parameters passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\description{
Plots a 2D simplex, a triangle with unit sides centered at the origin, onto
which per-recurrence posterior probabilities of recrudescence, relapse,
reinfection (or any other vector of three numbers in zero to one that sum to
one) can be projected; see \code{\link[=project2D]{project2D()}} and \strong{Examples} below.
}
\examples{
# Plot 2D simplex
plot_simplex(p.coords = diag(3),
             p.labels = c("(1,0,0)", "(0,1,0)", "(0,0,1)"),
             p.labels.pos = c(1,3,3))

# ==============================================================================
# Given data on an enrollment episode and a recurrence,
# compute the posterior probabilities of the 3Rs and plot the deviation of the
# posterior from the prior
# ==============================================================================

# Some data:
y <- list(list(m1 = c('a', 'b'), m2 = c('c', 'd')), # Enrollment episode
          list(m1 = c('a'), m2 = c('c'))) # Recurrent episode

# Some allele frequencies:
fs <- list(m1 = setNames(c(0.4, 0.6), c('a', 'b')),
           m2 = setNames(c(0.2, 0.8), c('c', 'd')))

# A vector of prior probabilities:
prior <- array(c(0.2, 0.3, 0.5), dim = c(1,3),
               dimnames = list(NULL, c("C", "L", "I")))

# Compute posterior probabilities
post <- compute_posterior(y, fs, prior)

# Plot simplex with the prior and posterior
plot_simplex(p.coords = rbind(prior, post$marg),
             p.labels = c("Prior", "Posterior"),
             pch = 20)

# Add the deviation between the prior and posterior: requires obtaining 2D
# coordinates manually
xy_prior <- project2D(as.vector(prior))
xy_post <- project2D(as.vector(post$marg))
arrows(x0 = xy_prior["x"], x1 = xy_post["x"],
       y0 = xy_prior["y"], y1 = xy_post["y"], length = 0.1)
}
