% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simplex.R
\name{plot_simplex}
\alias{plot_simplex}
\title{Plots a 2D simplex}
\usage{
plot_simplex(v_labels)
}
\arguments{
\item{v_labels}{A vector of labels with which to annotate vertices. If NULL,
vertices are not annotated.}
}
\description{
Plots a 2D simplex, a triangle with unit sides centered at the origin, to
which marginal posterior probabilities of the 3Rs can be added.
}
\examples{
# Plot 2D simplex
plot_simplex(v_labels = c("(1,0,0)", "(0,0,1)", "(0,1,0)"))

# ==============================================================================
# Given data on an enrollment episode and a recurrence,
# compute the posterior probabilities of the 3Rs and plot the deviation of the
# posterior from the prior
# ==============================================================================

# Some data:
y <- list(list(m1 = c("A", "C"), m2 = c("G", "T")), # Enrollment episode
          list(m1 = c("A"), m2 = c("G"))) # Recurrent episode

# Some allele frequencies:
fs <- list(m1 = setNames(c(0.4, 0.6), c("A", "C")),
           m2 = setNames(c(0.2, 0.8), c("G", "T")))

# A vector of prior probabilities:
prior <- array(c(0.2, 0.3, 0.5), dim = c(1,3),
               dimnames = list(NULL, c("C", "L", "I")))

# Compute posterior probabilities
post <- compute_posterior(y, fs, prior)

# Project marginal prior probabilities onto x and y coordinates:
xy_prior <- project2D(as.vector(prior))

# Project marginal posterior probabilities onto x and y coordinates:
xy_post <- project2D(as.vector(post$marg))

# Plot the deviation of the posterior from the prior
arrows(x0 = xy_prior["x"], x1 = xy_post["x"],
       y0 = xy_prior["y"], y1 = xy_post["y"], length = 0.1)
}
