% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pv3Rs-package.R
\docType{package}
\name{Pv3Rs-package}
\alias{Pv3Rs}
\alias{Pv3Rs-package}
\title{Pv3Rs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
