% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pv3Rs-package.R
\docType{package}
\name{Pv3Rs-package}
\alias{Pv3Rs}
\alias{Pv3Rs-package}
\title{Pv3Rs: Estimate the cause of recurrent vivax malaria using genetic data}
\description{
Estimate the per-person posterior probability that a recurrent Plasmodium vivax (Pv) infection is a recrudescence, relapse, reinfection (3Rs) using per-person P. vivax genetic data on more than one infection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aimeertaylor/Pv3Rs}
  \item Report bugs at \url{https://github.com/aimeertaylor/Pv3Rs/issues}
}

}
\author{
\strong{Maintainer}: Aimee Taylor \email{aimee.taylor@pasteur.fr} (\href{https://orcid.org/0000-0002-2337-8992}{ORCID})

Authors:
\itemize{
  \item Yong See Foo
}

Other contributors:
\itemize{
  \item European Union, Project 101110393 [funder]
}

}
\keyword{internal}
