% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pv3Rs-package.R
\docType{package}
\name{Pv3Rs-package}
\alias{Pv3Rs}
\alias{Pv3Rs-package}
\title{Pv3Rs: Estimate the Cause of Recurrent Vivax Malaria using Genetic Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Plot malaria parasite genetic data on two or more episodes. Compute per-person posterior probabilities that each Plasmodium vivax (Pv) recurrence is a recrudescence, relapse, or reinfection (3Rs) using per-person P. vivax genetic data on two or more episodes and a statistical model described in Taylor, Foo and White (2022) \doi{10.1101/2022.11.23.22282669}. Plot per-recurrence posterior probabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aimeertaylor.github.io/Pv3Rs/}
  \item \url{https://github.com/aimeertaylor/Pv3Rs}
  \item Report bugs at \url{https://github.com/aimeertaylor/Pv3Rs/issues}
}

}
\author{
\strong{Maintainer}: Aimee Taylor \email{aimee.taylor@pasteur.fr} (\href{https://orcid.org/0000-0002-2337-8992}{ORCID})

Authors:
\itemize{
  \item Yong See Foo (\href{https://orcid.org/0000-0003-3010-9106}{ORCID})
}

Other contributors:
\itemize{
  \item Tymoteusz Kwiecinski (compiled included macros.Rd) [compiler]
  \item Duncan Murdoch (author of included macros.Rd) [contributor]
  \item Mans Magnusson (author of included msg_progress_bar.R) [contributor]
  \item Institut Pasteur [copyright holder]
  \item European Union, Project 101110393 [funder]
}

}
\keyword{internal}
