% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pv3Rs-package.R
\docType{package}
\name{Pv3Rs-package}
\alias{Pv3Rs}
\alias{Pv3Rs-package}
\title{Pv3Rs: Estimate the Cause of Recurrent Vivax Malaria using Genetic Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compute per-person posterior probabilities that each recurrent Plasmodium vivax (Pv) recurrence is a recrudescence, relapse, or reinfection (3Rs) using per-person P. vivax genetic data on two or more episodes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aimeertaylor.github.io/Pv3Rs/}
  \item \url{https://github.com/aimeertaylor/Pv3Rs}
  \item Report bugs at \url{https://github.com/aimeertaylor/Pv3Rs/issues}
}

}
\author{
\strong{Maintainer}: Aimee Taylor \email{aimee.taylor@pasteur.fr} (\href{https://orcid.org/0000-0002-2337-8992}{ORCID})

Authors:
\itemize{
  \item Yong See Foo
}

Other contributors:
\itemize{
  \item Mans Magnusson [contributor]
  \item l'INSTITUT PASTEUR [copyright holder]
  \item European Union, Project 101110393 [funder]
}

}
\keyword{internal}
