% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG_to_igraph.R
\name{RG_to_igraph}
\alias{RG_to_igraph}
\title{Converts relationship graph to an \code{igraph} object}
\usage{
RG_to_igraph(RG, gs, ts_per_gs)
}
\arguments{
\item{RG}{Relationship graph output by \code{\link{enumerate_RGs_alt}} with
\code{igraph=FALSE}.}

\item{gs}{Vector of genotype names.}

\item{ts_per_gs}{Vector of infection numbers for each genotype.}
}
\value{
An \code{igraph} object along with the original variables in
\code{RG}.
}
\description{
Makes the relationship graph compatible with the output of
\code{\link{enumerate_RGs}}.
}
