% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lineages.R
\name{sample_lineages}
\alias{sample_lineages}
\title{Sample per-marker lineages for parasite genotypes in a relationship graph}
\usage{
sample_lineages(
  RG,
  n_m,
  outbred = TRUE,
  lineage_freqs = NULL,
  lineage_count = 50,
  equifrequent = TRUE,
  alpha = 1
)
}
\arguments{
\item{RG}{A relationship graph; see \code{\link{enumerate_RGs}} for details.}

\item{n_m}{Number of markers.}

\item{outbred}{Logical specifying if the parental genotypes should be sampled
from a outbred population or not.}

\item{lineage_freqs}{Vector of founder population lineage frequencies; if named,
names will be taken as lineages; only used if \code{outbred = FALSE}.}

\item{lineage_count}{Whole number specifying the number of lineages in the
founder population; only used if \code{outbred = FALSE} and
\code{lineage_freqs = NULL}.}

\item{equifrequent}{Logical specifying if the lineages in the population are
equifrequent or not; only used if \code{outbred = FALSE} and \code{lineage_freqs
  = NULL}.}

\item{alpha}{The concentration parameter of a symmetric Dirichlet
distribution over lineage frequencies when \code{equifrequent = FALSE}: set
\code{0 < alpha < 1} to sample sparse frequencies, \code{alpha = 1}
(default) to sample frequencies uniformly, and \code{alpha >> 1} to sample
frequencies close to \code{1/lineage_count}; only used if \code{outbred = FALSE},
\code{lineage_freqs = NULL} and \code{equifrequent = FALSE}.}
}
\description{
For each marker, the lineages of a set of first-generation parasite genotypes
are sampled, according to their inter-genotype relationships specified in a
relationship graph, from a set of parents sampled from a founder population
(a population in which each parasite genotype is represented by a single
lineage). The founder population can either be an infinitely
large outbred population or a finite inbred population. If its finite,
lineages can either be equifreqent or not. Markers are treated as
independent.
}
\details{
\code{sample_lineages} assume inter-genotype relationships are either
stranger, sibling or clone and samples lineages as follows.
\enumerate{

\item For a stranger, sample per-marker lineages from two parents.

\item For a clique of clones, sample per-marker lineages from two parents for
one member of the clique and clone it.

\item For a clique of siblings, sample per-marker lineages from two
parents.

\item For a mixed clique of siblings and clones, for each marker, from two
parents, sample a lineage for each clonal sub-clique and clone it.

}

Note that, since clonal lineages are drawn from two not necessarily identical
parents, clones are mitotic copies of one-another, not the products of
selfing.

When \code{outbred = TRUE} (default), this function assumes parasite
genotypes are drawn from an infinitely large outbred population. As such,
each parental lineage is unique. When \code{outbred = FALSE}, this function
assumes parasite genotypes are drawn from a finite inbred population. As
such, parental genotypes are not necessarily unique. Note that, by specifying
an inbred population representative of a brood of parasite genotypes, this
function could be used to explore IBD structure following brood mating.
}
\examples{
set.seed(3)
RG <- sample_RG(c(2, 1, 1))
plot_RG(RG)
sample_lineages(RG, n_m = 5)
sample_lineages(RG, n_m = 5, outbred = FALSE, lineage_count = 5)

}
