% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_RG.R
\name{sample_RG}
\alias{sample_RG}
\title{Sample a transitive relationship graph (RG)}
\usage{
sample_RG(MOIs)
}
\arguments{
\item{MOIs}{A numeric vector specifying, for each infection, the number of
distinct parasite genotypes, a.k.a the multiplicity of infection (MOI).}
}
\value{
Returns a transitive relationship graph (RG); see \code{\link{enumerate_RGs}} for details.
}
\description{
Samples one transitive graph of stranger and sibling relationships between
distinct parasite genotypes within an infection and stranger, sibling and
clonal relationships between parasite genotypes across infections. Unlike
\code{\link{enumerate_RGs}}, \code{sample_RG} is not limited to six or fewer
genotypes among three or fewer infections, because it only generates one
graph and thus cannot overload memory. However, for more than six genotypes
among more than three infections, it could take a long time because
\code{sample_RG} is testing for transitivity while sampling among not
necessarily transitive graphs, of which there are many; see the message
that \code{sample_RG} prints to the screen.
}
\details{
Superseded by \code{\link{sample_RG_alt}}.
}
\section{Provenance}{
 This function was adapted from
\code{generate_random_Vivax_model} at
\url{https://github.com/jwatowatson/RecurrentVivax/blob/master/Genetic_Model/iGraph_functions.R}.
}

\examples{
set.seed(2)
RG <- sample_RG(c(2, 1, 1))
print(RG)
igraph::as_data_frame(RG)
plot_RG(RG)

}
