% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ys_VHX_BPD}
\alias{ys_VHX_BPD}
\title{Example \emph{Plasmodium vivax} data}
\format{
A list of 217 study participants; for each study participant, a list of one or more
episodes; for each episode, a list of three or more microsatellite markers;
for each marker, a vector of observed alleles (repeat lengths). For example:
\describe{
\item{BPD_103}{Study participant identifier: study participant 103 in the BPD trial}
\item{BPD_103_1}{Episode identifier: episode one of study participant 103 in the BPD trial}
\item{PV.3.27}{Marker identifier: \emph{P. vivax} 3.27}
\item{18}{Allele identifier: 18 repeat lengths}
}
}
\source{
\itemize{
\item \url{https://github.com/jwatowatson/RecurrentVivax/blob/master/RData/GeneticModel/MS_data_PooledAnalysis.RData}
\item \url{https://github.com/aimeertaylor/Pv3Rs/blob/main/data-raw/ys_VHX_BPD.R}
}
}
\usage{
ys_VHX_BPD
}
\description{
Previously-published microsatellite data on \emph{P. vivax} parasites
extracted from study participants enrolled in the Best Primaquine Dose (BPD) and Vivax
History (VHX) trials; see
\href{https://www.nature.com/articles/s41467-019-13412-x}{Taylor & Watson et al. 2019}
for more details of the genetic data; for more details of the trials, see
\href{https://doi.org/10.1093/cid/ciy319}{Chu et al. 2018a} and
\href{https://doi.org/10.1093/cid/ciy735}{Chu et al. 2018b}.
}
\keyword{datasets}
